ai/quiz/requirements.txt
openai==1.14.3
mysql-connector-python==8.3.0
python-dotenv==1.0.1
